Correctly receiving a directory

+++ app/src/main/java/com/xtech/sultano/optimizedfilesender/DownloadArrayAdapter.java
@@ -1,6 +1,7 @@
 package com.xtech.sultano.optimizedfilesender;
 
 import android.content.Context;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -83,8 +84,8 @@ public class DownloadArrayAdapter extends ArrayAdapter<Download> {
         //Set the progress Status
         TextView progressBarText = (TextView) v.findViewById(R.id.download_progressbar_label);
 
-        float fileSize = d.getDimension() / 1024 / 1024;
-        float sentData = d.getSentData() / 1024 / 1024;
+        float fileSize = (float)d.getDimension() / 1024 / 1024;
+        float sentData = (float)d.getSentData() / 1024 / 1024;
         progressBarText.setText(String.format("%d%% - %.2f / %.2f Mb", d.getProgress(), sentData, fileSize));
 
 /*        ProgressBar progressBar = (ProgressBar)v.findViewById(R.id.send_progress_bar);
